requirements.txt
requests
beautifulsoup4
psycopg2
python-dotenv
cloudscraper
pika
